# docs/swagger.yaml
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /songText/{id}:
    get:
      consumes:
      - application/json
      description: Получает текст песни по её ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить текст песни по ID
      tags:
      - songs
...
swagger: "2.0"